+++ app/src/main/java/com/kuit/ourmenu/ui/onboarding/viewmodel/LoginViewModel.kt
@@ -4,8 +4,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kuit.ourmenu.data.model.auth.SignInType
 import com.kuit.ourmenu.data.repository.AuthRepository
-import com.kuit.ourmenu.ui.oauth.KakaoModule.getUserEmail
-import com.kuit.ourmenu.ui.onboarding.state.KakaoState
 import com.kuit.ourmenu.ui.onboarding.state.LoginState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
@@ -28,9 +26,6 @@ class LoginViewModel @Inject constructor(
     private val _loginState: MutableStateFlow<LoginState> = MutableStateFlow(LoginState.Default)
     val loginState = _loginState.asStateFlow()
 
-    private val _kakaoState = MutableStateFlow<KakaoState>(KakaoState.Default)
-    val kakaoState = _kakaoState.asStateFlow()
-
     private val _error: MutableStateFlow<String?> = MutableStateFlow(null)
     val error = _error.asStateFlow()
 
@@ -40,50 +35,6 @@ class LoginViewModel @Inject constructor(
 
     fun updatePassword(password: String) {
         _password.value = password
-    }
-
-    fun signInWithKakao() {
-        viewModelScope.launch {
-            val email = getUserEmail()
-            email ?: run {
-                _error.value = "이메일을 가져오는데 실패했습니다."
-                return@launch
-            }
-            _kakaoState.value = KakaoState.Loading
-
-            if (checkKakaoEmail()) {
-                authRepository.login(
-                    email = email,
-                    password = null,
-                    signInType = SignInType.KAKAO
-                ).fold(
-                    onSuccess = {
-                        _kakaoState.value = KakaoState.Login
-                    },
-                    onFailure = { error ->
-                        _kakaoState.value = KakaoState.Error
-                        _error.value = error.message
-
-                        delay(1000)
-                    }
-                )
-            } else {
-                // 카카오 회원 X -> 회원가입
-                _kakaoState.value = KakaoState.Signup
-            }
-
-        }
-    }
-
-    private suspend fun checkKakaoEmail(): Boolean {
-        val email = getUserEmail()
-
-        authRepository.checkKakaoEmail(email).fold(
-            onSuccess = { response ->
-                return response?.existUser == true
-            },
-            onFailure = { return false }
-        )
     }
 
     fun signInWithEmail() {
